+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ### Fixed
 - \#2691 - Filters don't updated correctly on reload
+- \#2699 - App list health bar update/render issue
 
 ## 0.13.4 - 2015-11-18
 ### Fixed